+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes:
 
 * Added support for dark theme mode on Windows (including windows title bar background),
   MacOS and GNU/Linux.
+* Dataset edit layout: now disabling/enabling "Apply" button depending on widget value changes
 
 ### Version 1.8.0 ###